Implement INotifyPropertyChanged for these
Private implementation uses int to avoid issues with enums and generic methods.
Unsafe.As is the fastest way to convert and is the same operation happening under the hood anyway even when it's an enum both places, but enum involves additional code that is guaranteed not needed here anyway, so this is a win.